changed data sum of squares to double

+++ fast_matched_filter/src/matched_filter.c
@@ -28,10 +28,10 @@ void matched_filter(float *templates, float *sum_square_templates, int *moveouts
     int network_offset, station_offset, cc_sum_offset;
     int *moveouts_t = NULL;
     float *templates_t = NULL, *sum_square_templates_t = NULL, *weights_t = NULL;
-    float *csum_square_data = NULL;
+    double *csum_square_data = NULL;
 
     // compute cumulative sum of squares of data
-    csum_square_data = malloc((n_samples_data * n_stations * n_components + 1) * sizeof(float));
+    csum_square_data = malloc((n_samples_data * n_stations * n_components + 1) * sizeof(double));
     csum_square_data[0] = 0.0;
     cumsum_square_data(data, n_samples_data, weights, n_stations, n_components, csum_square_data + 1);
 
@@ -78,7 +78,7 @@ void matched_filter(float *templates, float *sum_square_templates, int *moveouts
  
 //-------------------------------------------------------------------------
 float network_corr(float *templates, float *sum_square_template, int *moveouts,
-                   float *data, float *csum_square_data, float *weights,
+                   float *data, double *csum_square_data, float *weights,
                    int n_samples_template, int n_samples_data, int n_stations, int n_components) {
 
     int s, c, d, dd, t;
@@ -111,7 +111,7 @@ float network_corr(float *templates, float *sum_square_template, int *moveouts,
  
 //-------------------------------------------------------------------------
 float corrc(float *templates, float sum_square_template,
-            float *data, float *csum_square_data,
+            float *data, double *csum_square_data,
             int n_samples_template) {
 
     int i;
@@ -120,7 +120,7 @@ float corrc(float *templates, float sum_square_template,
     for (i = 0; i < n_samples_template; i++){
         numerator += templates[i] * data[i];
     }
-    denominator = sum_square_template * (csum_square_data[n_samples_template] - csum_square_data[-1]);
+    denominator = sum_square_template * (float)(csum_square_data[n_samples_template] - csum_square_data[-1]);
 
     if (denominator > STABILITY_THRESHOLD) cc = numerator / sqrt(denominator);
 
@@ -130,7 +130,7 @@ float corrc(float *templates, float sum_square_template,
 //-------------------------------------------------------------------------
 void cumsum_square_data(float *data, int n_samples_data, float *weights,
                         int n_stations, int n_components,
-                        float *csum_square_data) {
+                        double *csum_square_data) {
     int s, c;
     int station_offset, component_offset, d;
 
@@ -149,15 +149,15 @@ void cumsum_square_data(float *data, int n_samples_data, float *weights,
 }
 
 //-------------------------------------------------------------------------
-void neumaier_cumsum_squared(float *array, int length, float *cumsum) {
+void neumaier_cumsum_squared(float *array, int length, double *cumsum) {
     int i;
-    float running_sum, square, temporary;
-    float correction = 0.0;
+    double running_sum, square, temporary;
+    double correction = 0.0;
 
-    running_sum = array[0] * array[0];
+    running_sum = (double)array[0] * (double)array[0];
     cumsum[0] = running_sum;
     for (i = 1; i < length; i++) {
-        square = array[i] * array[i];
+        square = (double)array[i] * (double)array[i];
         temporary = running_sum + square;
 
         if (fabsf(running_sum) >= fabsf(square)) {
